traitement_data/requirements.txt
pandas>=2.0.0
pyarrow>=12.0.0
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.22.0
opencv-python>=4.7.0
scikit-learn>=1.2.0
datasets>=2.0.0
scipy>=1.10.0
